Fix get_whisper_prompt() iteration
Fix iteration so it correctly adds all data examples to the prompt. It
iterates first on examples, then on variables, so the resulting list
follows the structure: "color, red. weight, 70. color, blue. weight, 72"
instead of the previous "color, red. color, blue. weight, 70. weight,
72".

This older design sometimes led Whisper to made incorrect repetitions
of the last variable in the prompt. In addition, the new structure is
closer to real world audio instructions.